labs/lab2/wealth-advisor-agent/pyproject.toml
[tool.poetry]
name = "wealth-advisor-agent"
version = "0.1.0"
description = "A crewAI agent."
authors = ["Your name <email-id>"]
license = "MIT"
readme = "README.md"
packages = [
    { include = "assistance_crew", from = "." }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
crewai = "^0.102"
ibm-watsonx-ai = "^1.3.3"
tavily-python = "^0.5.3"

[tool.poetry.group.dev]
optional = true

[tool.poetry.group.dev.dependencies]

python-dotenv = "^1.0.1"
ibm-watsonx-ai = "^1.2.7"
pytest = "^8.3.3"
numpy = "<2.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
